Rename packages to match styling
