+++ source/core/src/main/com/csse3200/game/screens/AirAnimalSelectionScreen1.java
@@ -6,7 +6,7 @@ import com.csse3200.game.GdxGame;
 import com.csse3200.game.components.animal.AirAnimalSelectionDisplay;
 import com.csse3200.game.components.animal.AnimalRouletteActions1;
 import com.csse3200.game.components.animal.AnimalRouletteDisplay1;
-import com.csse3200.game.ui.PopUpDialogBox.PopUpHelper;
+import com.csse3200.game.ui.pop_up_dialog_box.PopUpHelper;
 
 public class AirAnimalSelectionScreen1 extends AnimalRouletteScreen1 {
     public AirAnimalSelectionScreen1(GdxGame game) {
