+++ source/core/src/main/com/csse3200/game/screens/AnimalRouletteScreen1.java
@@ -13,7 +13,7 @@ import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.csse3200.game.GdxGame;
 import com.csse3200.game.components.animal.AnimalRouletteActions1;
 import com.csse3200.game.components.animal.AnimalRouletteDisplay1;
-import com.csse3200.game.ui.PopUpDialogBox.PopUpHelper;
+import com.csse3200.game.ui.pop_up_dialog_box.PopUpHelper;
 
 public abstract class AnimalRouletteScreen1 extends ScreenAdapter {
     protected Stage stage;
