+++ source/core/src/main/com/csse3200/game/ui/pop_up_dialog_box/PopupDialogBox.java
@@ -1,4 +1,4 @@
-package com.csse3200.game.ui.PopUpDialogBox;
+package com.csse3200.game.ui.pop_up_dialog_box;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
